Insert investor data in Person table

+++ api-investor/InvesterSystem.DataModel/Table/MetaData/Gender.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -11,7 +12,7 @@ namespace InvestorSystem.DataModel.Table
     {
         [Key]
         [MaxLength(1)]
-        public int ID { get; set; }
+        public short ID { get; set; }
 
         public string Name { get; set; }
